Nodes account for what runs, not what once ran: a reservation ledger, fractional CPU, and foreground cleanup (#588)
* Nodes account for what runs, not what once ran: a reservation ledger, fractional CPU, and foreground cleanup

A node reserved CPU, memory and disk when a workload started and gave them back only when
it was destroyed, so every finished sync and publish VM kept a whole core and its memory until
the next deploy retired it, and two deployed projects filled a 4-core node. The node record now
keeps a ledger of the room each workload holds, keyed by workload id: a run holds everything,
an ended run holds its disk alone, a restart takes CPU and memory again and fails when the node
lacks them, and reserving or releasing the same workload twice changes nothing. Registration
rebuilds the ledger from the workload records, so a node that re-registers stops carrying a
leak forward.

Workload.vcpus becomes cpus, a double: a fraction of a core is a CPU quota on the VM, and the
node's availableCpus moves by the same fraction. The sync VM takes 1 core, a runtime VM and
the site workloads 0.5; the sync disk drops to 2 GB and the runtime disk to 512 MB.

A foreground workload is destroyed by the orchestrator as soon as its reply arrives, whatever
its outcome, including a start that failed: its VM, its room and its record are gone. Its logs
stay in Loki, so LogService reads by organization and workload id instead of through the
workload record, the way TelemetryService already does, and the consoles pass the organization
along.


* Relock the frontend on the published client packages


* The VM sizes that never vary per environment are constants, not properties

syncCpus, runtimeCpus, syncDiskSizeMb and runtimeDiskSizeMb move from DeploymentProperties to
ProjectWorkloadSizes; the memory sizes and the checkout mount limit stay configurable.


* The sync VM takes half a core too


* The deployment VM memory sizes and checkout mount cap are constants too


* A run that has ended leaves nothing on the node: the vm-manager removes its VM, its record is its history

Every run holds its whole reservation until it ends, and nothing is kept for a restart in
place: once the server holds a run's outcome, the node ships its logs, removes the VM, its disk
and its log files, and forgets it. The workload record stays on the server with the status and
exit code, and its logs stay in Loki under its id. Workload.autoRemove goes, being always on;
restartWorkload goes, there being no VM to boot; a microservice restart from the console runs
a fresh VM from the project's current deployment through the new ProjectWorkloadFactory, which
the deploy job shares. The deploy no longer retires the previous run's workloads or destroys an
ended runtime before replacing it, and the ledger loses its running flag.


* The schema change is a new migration, and the node's free capacity gets its own columns

V1 and V5 are back as the running environments applied them. V6 adds the ledger and the
fractional CPU columns: Elasticsearch cannot retype a mapped field, so kinotic_vm_node's
INTEGER available* columns stay mapped and unused and freeCpus/freeMemoryMb/freeDiskMb
replace them, filled when the node registers; kinotic_workload gains cpus beside the unused
vcpus and autoRemove. V7 sizes the console fixtures in the new columns.


* Deregistering a node counts the workloads still running on it, not its history


* A microservice deployment reads FAILED once its VM has ended

The row records the outcome of deploying and the workload the outcome of running; the listing
folds the run's end in, so a VM that died no longer shows DEPLOYED until the next push.


---------

+++ kinotic-system-api/src/main/java/org/kinotic/system/api/services/VmNodeOrchestrationService.java
@@ -77,11 +77,11 @@ public interface VmNodeOrchestrationService {
     /**
      * Finds a node with sufficient resources to host a workload with the given requirements.
      *
-     * @param requiredCpus the number of vCPUs required
+     * @param requiredCpus the CPU required, in cores
      * @param requiredMemoryMb the amount of memory required in megabytes
      * @param requiredDiskMb the amount of disk space required in megabytes
      * @return a future that will complete with a suitable node, or null if none available
      */
-    Future<VmNode> findAvailableNode(int requiredCpus, int requiredMemoryMb, int requiredDiskMb);
+    Future<VmNode> findAvailableNode(double requiredCpus, int requiredMemoryMb, int requiredDiskMb);
 
 }
